feat: support ollama agent
